Update gitignore and add QM7 split indices
- Add TensorBoard exclusions to .gitignore
- Add QM7 dataset split indices
- Remove old pretrained checkpoints
- Clean up swap files

+++ bnn_aenet/datamodule/aenet/data_loader.py
@@ -355,7 +355,7 @@ class PrepDataloader(object):
 
 				aux_iesp.append(aux)
 
-			aux_iesp = torch.tensor( np.array(aux_iesp) )  
+			aux_iesp = torch.tensor( np.array(aux_iesp), dtype=torch.float32 )  
 			logic_tensor_reduce.append(aux_iesp)
